Implemented changes based on PR comments

+++ src/modules/Elsa.Workflows.Api/Endpoints/WorkflowDefinitions/DeleteVersion/Endpoint.cs
@@ -2,6 +2,7 @@ using Elsa.Abstractions;
 using Elsa.Workflows.Api.Constants;
 using Elsa.Workflows.Management.Contracts;
 using JetBrains.Annotations;
+using Microsoft.AspNetCore.Authorization;
 
 namespace Elsa.Workflows.Api.Endpoints.WorkflowDefinitions.DeleteVersion;
 
@@ -9,21 +10,30 @@ namespace Elsa.Workflows.Api.Endpoints.WorkflowDefinitions.DeleteVersion;
 internal class DeleteVersion : ElsaEndpoint<Request>
 {
     private readonly IWorkflowDefinitionManager _workflowDefinitionManager;
+    private readonly IAuthorizationService _authorizationService;
 
-    public DeleteVersion(IWorkflowDefinitionManager workflowDefinitionManager)
+    public DeleteVersion(IWorkflowDefinitionManager workflowDefinitionManager, IAuthorizationService authorizationService)
     {
         _workflowDefinitionManager = workflowDefinitionManager;
+        _authorizationService = authorizationService;
     }
 
     public override void Configure()
     {
         Delete("/workflow-definition-versions/{id}");
         ConfigurePermissions("delete:workflow-definitions");
-        Policies(AuthorizationPolicies.NotReadOnlyPolicy);
     }
 
     public override async Task HandleAsync(Request request, CancellationToken cancellationToken)
     {
+        var authorizationResult = _authorizationService.AuthorizeAsync(User, null, AuthorizationPolicies.NotReadOnlyPolicy);
+
+        if (!authorizationResult.Result.Succeeded)
+        {
+            await SendForbiddenAsync(cancellationToken);
+            return;
+        }
+
         var deleted = await _workflowDefinitionManager.DeleteByIdAsync(request.Id, cancellationToken);
 
         if (!deleted)
